Resolve issue http://jira.jahia.org/browse/JAHIA-3689 Page field: impossible to create an empty link_only page field: title is removed
Actually, the option to let the user "not create a page or link" has been removed, because it does not work. Indeed, you cannot create a container without creating a link, but still display the title you entered. You will get a N.D. and not the title you entered. 

Also, the utility of such an option is a bit strange. If you do not want to create a page or a link, then just skip this field.

git-svn-id: svn://localhost/jahia/trunk@93 69c369ed-ed09-0410-a4a1-e19a46c23626