Brand new profile available to configure Jahia BEFORE packaging. This way you can now simply add the following to your settings.xml file :
<activeProfile>configure-before-packaging</activeProfile><activeProfile>configure-before-packaging</activeProfile>

and when building the WAR file Jahia will be configured before-hand. You will therefore get a fully configured, ready-to-deploy WAR and EAR files. This should make deployment to JEE application servers a lot more portable.

In the long term this will be part of the configuration wizard that will possibly embed Maven.

git-svn-id: svn://localhost/jahia/trunk@153 69c369ed-ed09-0410-a4a1-e19a46c23626