TEMPLATES-659: Mix of resource bundles within the same page: not pointing to the same resources http://jira.jahia.org/browse/TEMPLATES-659 http://10.8.37.247/fisheye/cru/AND-Wroom-57
Cause of the problem:
This is a normal ResourceBunlde behaviour.
The lookup for a bundle is made in the following sequence:
http://java.sun.com/j2se/1.4.2/docs/api/java/util/ResourceBundle.html#getBundle(java.lang.String,%20java.util.Locale,%20java.lang.ClassLoader)
For example, I am looking for message.properties in it_IT. The default JVM's locale is fr. The lookup in this case is:
 - message_it_IT.properties
 - message_it.properties
 - message_fr.properties
 - message.properties

The JSTL makes it in a different way, as they consider the usage of the default JVM's locale as a bad design.


Resolution:
I've added a parameter into jahia.properties:

considerDefaultJVMLocal                                =   false

By default, we ignore now the default JVM's locale (Locale.getDefault()) and use the bundle with the basename (no locale) if the one for language is not found.
Setting considerDefaultJVMLocal to "true" will switch back the old behaviour, but I am not in favour of it, as the behaviour of the ResourceBundle in this case depends on the installation. I.e., if Jahia is installed on a machine with the French default locale, the missing labels will be in French. If the machine has German as default (my case), the missing labels are in German. This is definitely not what the template developers want.

git-svn-id: svn://localhost/jahia/branches/JAHIA-6-0-BRANCH@1698 69c369ed-ed09-0410-a4a1-e19a46c23626