Small commit to reduce the usage of ProcessingContext, ParamBean and JahiaData in Jahia :)
Actually this should minimize the exposure of all the org.jahia.data.* objects in the templates, so that template developers are not tempted to use them.

Some tags were modified to use the RenderContext and it's relevant objects instead, while others will still look for JahiaData only if RenderContext is not available. This is needed because some tags are used both in templates and the administration.

git-svn-id: svn://localhost/jahia/trunk@4533 69c369ed-ed09-0410-a4a1-e19a46c23626