QA-1452
Remove content tab from site manager as this one is only listing virtual site and pages and is made for reordering/deleting pages, etc.

Actually this site manager seems to not make much sense anymore (as we have same functions directly on the left panel of the edit mode), so I have remove this tab.

If we find a real usage and need to display once again the content tab it is not an issue it is a simple line of code in the xml file (we cans also hide it for virtaul site for example and keep it for pages we have choice).

git-svn-id: svn://localhost/jahia/trunk@9653 69c369ed-ed09-0410-a4a1-e19a46c23626