NGA-10: Partial published site throws error http://support4.jahia.com/browse/NGA-10 http://10.8.37.247/fisheye/cru/AND-134
Problem: in the NGA-10 the published "Activities" page uses absolute modules from the home page, which is not published. Thus the home page cannot be found and we result in an NPE and "Module error" on the rendered page.

Resolution: correctly check of the node (where the absolute area is coming from) can be resolved

git-svn-id: svn://localhost/jahia/trunk@12990 69c369ed-ed09-0410-a4a1-e19a46c23626