QA-3093 Drag & Drop bugfixes
This commit fixes a few issues. Now when we have named child definitions, we :
 - Generate the proper green buttons
 - The green button are click-able and propose the proper types
 - If a reference type is not allowed as a target drop, it will not work

So you can now define a node as such : 

jnt:myobject > nt:base 
+ first (jnt:content) 
+ second (jnt:content) 

This patch was however NOT tested with a definition such as : 

jnt:myobject > nt:base 
+ first (jnt:content) 
+ second (jnt:content) 
+ * (jnt:content)

This should be done to see how this behaves, as it might generate another ticket (I suggest we open another one for this case).

git-svn-id: svn://localhost/jahia/branches/JAHIA-6-6-X-X-BRANCH@15160 69c369ed-ed09-0410-a4a1-e19a46c23626