BACKLOG-2253 - RB improvements http://jira.jahia.org/browse/BACKLOG-2253
Make native2ascii task easier to configure (move the most part to the parent jahia-modules).
Added support for creating the default-locale resource bundle if not present (the locale of the default bundle is configurable and is set by default to "en").

TODO: adjust the jahia-modules pom.xml and sub-module configuration

git-svn-id: svn://localhost/jahia/branches/JAHIA-6-6-X-X-BRANCH@15226 69c369ed-ed09-0410-a4a1-e19a46c23626