New JSON view to output a node's type information, you can either retrieve it with the following : - node.nodetype.json - node.full.json?withType=true
There are also two optional parameters that allow to control output of hidden and protected type information. By default the hidden and protected property definition and child definitions will not be output, but you can ask to generate them by setting the withHidden=true or withProtected=true query string parameters.

So if you want all the node data, the first level of childs, as well as the node's type including hidden and protected properties and child definitions, you can use an url of the form : 

- node.full.json?withType=true&withHidden=true&withProtected=true (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 41914)

git-svn-id: svn://localhost/jahia/trunk@16680 69c369ed-ed09-0410-a4a1-e19a46c23626