BOUYGUES-47 : Add logging and servlet context mime type resolution in the case of an invalid or legacy file structure being imported.
In the client ticket, they had manually created an invalid file structure that was still being recognized as valid as it respected an old export format we had. Unfortunately in the legacy format the mimetype was on the file node instead of the jcr:content node so the import couldn't find the mimetype anymore. This ticket logs the problem so that we can more easily detect it again if another customer tries to do the same thing, and also uses the ServletContext mime type resolution to avoid importing a wrong mime type.

Another solution would be to remove the jcr:mimeType from the list of excluded attributes in the applicationcontext-jcr.xml file but the potential side effects of this were not easy to determine so this patch was created instead. (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 42140)

git-svn-id: svn://localhost/jahia/trunk@16906 69c369ed-ed09-0410-a4a1-e19a46c23626