BOUYGUES-67 & BOUYGUES-70 OutOfMemory while exporting site
This patch modifies the Jahia access manager to no longer used an unbounded map to store path and permission cache, but instead uses a limited (but configurable in jahia.properties) LRU map. 

This was identified as a problem on large repositories where operations that check permissions on large amounts of nodes can grow this cache structure quite fast, and in the case of the customer, it grew to over 1.5GB in RAM ! So this patch will help limit the size.

It is a tradeoff with performance, since if these values are needed often and the cache is too small, we will have to re-evaluate the permissions more often, so the default value of the sizing should be tested with performance and volumetry testing. The default value of 1000 was arbitrary at this point.

To test this, I recommend using a large site and testing an export with a YourKit CPU profiling (no J2EE profiling activated please) activated and making sure that the behavior of the application is still normal. (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 43155)

git-svn-id: svn://localhost/jahia/trunk@17921 69c369ed-ed09-0410-a4a1-e19a46c23626