CGBR-49 Add missing portlet validator
This change adds a new import validator that checks if the import file is referencing any portlets that have not been deployed, defined, or that are referenced from other sites that have not been imported.

In order to test this, perform the following scenario:

1. Create a site
2. Deploy a portlet
3. Create a portlet instance inside the site
4. Use the portlet instance in a page
5. Publish the site
6. Export the site
7. Install a new Jahia, do NOT deploy the portlet
8. Try to import the site, it should fail, indicating in the logs that a portlet is missing, such as in the following example : 

2012-11-01 17:07:17,481: ERROR [ManageSites] - Failed Import missingportlet.zip/live-repository.xml validated in 46629 ms: [overall result=failure, details=[[MissingModulesValidationResult=successful], [MissingTemplatesValidationResult=successful, targetTemplateSet=templates-web-space, targetTemplateSetPresent=true, templateSetsMissingCounts={}, missingTemplates={}], [MissingNodetypesValidationResult=successful], [MissingPortletsValidationResult=failure, missingPortlets=[/sites/missingportlet/portlets/Test_x0020_Portlet_x0020

You can also test by using inside a site a portlet instance that was defined in another site that was then not imported on the new Jahia instance. This should also fail. (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 43554)

git-svn-id: svn://localhost/jahia/trunk@18320 69c369ed-ed09-0410-a4a1-e19a46c23626