Do not search in the index for hierarchical facets to be updated in case of the version index.
I saw in profiler snapshots when working on the version clean tool ( https://jira.jahia.org/browse/BACKLOG-3770 ) that when version nodes are deleted and thus removed from the index the code is triggered to get a read-only index reader and search over it in order to check for possible facet-related nodes. On the version store index this can be skipped as no facet-related info is indexed there. (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 45337)

git-svn-id: svn://localhost/jahia/trunk@20103 69c369ed-ed09-0410-a4a1-e19a46c23626