HOME-90 QA-4210 Bad performance with LDAP
In this modification I am deactivating by default the check on the global groups as it seemed, after discussions with Thomas and Sergiy, that we don't really need it. In order to be able to react quickly should this change break anything, it has been made configurable so it can be quickly reactivated using a jahia.properties variable. (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 45595)

git-svn-id: svn://localhost/jahia/trunk@20361 69c369ed-ed09-0410-a4a1-e19a46c23626