BOUYGUES-109 Problem with multiple click on save button
This fix simply deactivated the Save button while the remote procedure call is executing and reactivates it to it's previous state when the call is completed (either successfully or not).

To test, simply try to quickly click the Save button twice when editing an ACL. Before this change it will execute the save twice, after this change you should not be able to click on it twice since it is disabled.

git-svn-id: svn://localhost/jahia/branches/JAHIA-6-6-X-X-BRANCH@20842 69c369ed-ed09-0410-a4a1-e19a46c23626