QA-4257 Pagination URL parameters are not properly encoded
This was a general problem in our extension of the c:url tag implementation because we forgot to encode the parameters we retrieve from the request map. This fix actually encodes all the parameters that have been retrieved from the request map exactly the same way that we do it for the c:param tag.

git-svn-id: svn://localhost/jahia/branches/JAHIA-6-6-X-X-BRANCH@20869 69c369ed-ed09-0410-a4a1-e19a46c23626