BACKLOG-3979 - Allows to set distribution management entry from the studio https://jira.jahia.org/browse/BACKLOG-3979
BACKLOG-4119 - extend release module dialog box
https://jira.jahia.org/browse/BACKLOG-4119

BACKLOG-3975 - Define the link from maven server to forge server
https://jira.jahia.org/browse/BACKLOG-3975

Allow to provide distribution management server info for a more (before releasing it).
Check the target distribution repository for a "link" to a Jahia Catalog server (Forge), by checking for a file jahia-catalog.properties in the root of the repository. The file contains a URL to the forge server.
Exetnd the release module dialog with options to publish artefacts to Maven and Jahia Catalog and also additional parameters for the Catalog (credentials, etc.)

git-svn-id: svn://localhost/jahia/trunk@21493 69c369ed-ed09-0410-a4a1-e19a46c23626