QA-4711 When Channels activated and no device specified => NPE, leads to empty zone in page
This patch will simply log a warning indicating that the channel is not properly configured and render the content normally

To reproduce the issue, simply go in edit mode on a content object, check the "Select channels" checkbox in the channels tab but DON'T select any channels and click ok. Publish the page and go into live mode and you should get the exception reported in the QA ticket.

This patch will change this into a logging statement that looks like this : 

2013-11-08 10:12:03,534: WARN  [ChannelExclusionFilter] - Channel selection activated on resource /sites/ACME-SPACE/home/activities/maincontent/acme-space-activities-in-brief..html but no channels selected, please select channels for this function to be activated properly. (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 47848)

git-svn-id: svn://localhost/jahia/trunk@22614 69c369ed-ed09-0410-a4a1-e19a46c23626