QA-4709 - j:extractedText unnecessarily occupies much memory https://jira.jahia.org/browse/QA-4709
Convert j:extractedText from string to binary to use the data store for it.
This reduces memory (Jackrabbit caches), as the content of that binary property is loaded only on demand, unless the length of that binary is less than minRecordLength bytes (1024 by default), in which case the content is stored as a byte array together with the node.
Also one benefit is when a PDF (with the j:extractedText) is published the j:extractedText is still references to the same "j:extractedText"-file entry in the data store and not gets duplicated.
On the opposite side, the file content is read when re-indexing the file node with the j:extractedText property.


git-svn-id: svn://localhost/jahia/trunk@22651 69c369ed-ed09-0410-a4a1-e19a46c23626