UPGR-9 Startup issue with jahia.com installation
This fixes the startup issue my modifying the behavior of the LastModifiedListener to detect moves or renames of content and to avoid auto-publishing such nodes. 

As this is a modification in a rather core component, it might have side effects, although every effort was made to keep them minimal. Please review carefully.

For testing, I suggest that the following features be tested to make sure they still work (please also test before this change as they might not all work already :))
- Node add & publishing
- Page move & publishing
- Node rename & publishing
- Node reordering among childs & publishing
- Node deleting & publishing
- Import content & publishing


git-svn-id: svn://localhost/jahia/trunk@23499 69c369ed-ed09-0410-a4a1-e19a46c23626