Fixed merge for: EUROPCAR-46 - Uniqueness of vanity url https://support.jahia.com/browse/EUROPCAR-46
Check uniqueness of vanity URL mapping for both workspaces.
Perform the check with system session and not with the current user's session (to avoid cases where the current user cannot query all the nodes in the site and thus the check won't be complete).
Use i18n message for the error message.
Do not log exception as this is just a normal validation and not an error. (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 48920)

git-svn-id: svn://localhost/jahia/trunk@23687 69c369ed-ed09-0410-a4a1-e19a46c23626