CRLO-91 Fix namespace register null prefix
In this problem we are suspecting that a null prefix is inserted because of some kind of concurrent modification of the hashmap since it is not synchronized. This seems like the only plausible explanation since we do check for null prefix and URI when registering namespaces so this shouldn't be possible otherwise.

Also this fix correct the incorrect messages in the getPrefix and getUri methods.

Note that I am never fond of adding synchronization as it might always have a performance impact, but this should only occur when registering namespaces, which shouldn't take much time and shouldn't happen too often in a stable environment.

We have no reproduction scenario for this problem, but it seems that they have some portlets deployed that seem to cause this problem at startup. So maybe testing with mutiple deployed portlets might be a good test. (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 49281)

git-svn-id: svn://localhost/jahia/trunk@24047 69c369ed-ed09-0410-a4a1-e19a46c23626