SOC-264 invalid encoding in file reference
This issue is actually due to the fact that when using GWT content rendering we do not set the response character encoding (while we do it in Render.doGet).

This patch sets the GWT content render response to the same value of the Render.doGet method.

To test : 
1. Create a file reference with a title such as : ?\208?\161?\209?\130?\209?\128?\208?\176?\208?\189?\208?\184?\209?\134?\208?\176 ?\208?\186?\208?\176?\209?\130?\208?\176?\208?\187?\208?\190?\208?\179?\208?\176 (make sure you deactivate system name synching otherwise it will generate a NullPointerException but this is another issue and quite minor) 
2. Upon refresh, the title should be properly displayed after this patch, and not before.

Note that this problem only occurs for pages loaded from the GWT navigation. If the page is loaded directly from the browser URL even in edit mode the problem does not occur, since in that case the page is loaded through the Render servlet and not through the GWT getRenderedContent AJAX call.

Also note that Jahia 7 is NOT affected by this bug since we do not load the pages in edit mode through an AJAX getRenderContent call anymore but instead load inside an iframe. (merged from jahia/branches/JAHIA-6-6-X-X-BRANCH revision 49694)

git-svn-id: svn://localhost/jahia/trunk@24460 69c369ed-ed09-0410-a4a1-e19a46c23626