BACKLOG-5207 User object is now not updated in the session
- We now update the user in the session in the case where no login or logout is done as was before all these changes
- In the case of the login we put the user in the session
- In the case of the logout we DON'T put the user in back in the session (this is what we wanted to fix)

It is *possible* that the setAttribute call generates an IllegalStateException is the session was invalidated just before the method call. It is unknown at this point how often this can occur and QA testing (with Selenium) to determine if these exceptions are frequent or not. On basic testing they didn't occur, but if they are a problem we can catch them to log the error in a less violent way :) 

This was already discussed with Benjamin btw. (merged from jahia/branches/JAHIA-7-0-X-X-BRANCH revision 50858)

git-svn-id: svn://localhost/jahia/trunk@25624 69c369ed-ed09-0410-a4a1-e19a46c23626