QA-6810 Webdav servlet exposes sensitive data
To fix this we simply retrieve the property twice, once to load it from the back-end, and the second time from the parent node (which is only possible once we loaded the property from the back-end), to make sure that any property filtering is indeed processed.

Fixing this problem also exposed some issues in the JCRNodeDecorator implementation that wasn't properly creating property wrapper that reference the proper parent node instance. This fix also includes a fix for this.

To test, simply try to request this URL :  
http://localhost:8080/server/live/jcr:root/users/root/j:password  

Before the fix you should get the property, after the fix you will get a 404.

I also recommend testing this modification at a performance and memory consumption level, as it may have impacts since we are creating/destroying more JCRPropertyWrapperImpl but it is difficult to know from the code how much of a problem this might actually be.

This fix also contains a correction for the JCRNodeWrapperImpl.setProperty(name, Value[]) methods that didn't properly handle properties removed by setting the value to null as well as an implementation for the JCRNodeWrapperImpl.getProperties(String[] namePatterns) method that wasn't implemented. (commit by shuber merged from jahia/branches/JAHIA-7-0-X-X-BRANCH revision 51380)

git-svn-id: svn://localhost/jahia/trunk@26146 69c369ed-ed09-0410-a4a1-e19a46c23626