QA-8909: Stabilize hashCode of Resource object
The problem was that if Resource.template is null
Resource.getResolvedTemplate will compute a template, but when restoring
the Resource from cacheKey that computed template is put directly to
Resource.template. And with that the same Resource could have two
different hashcodes resulting in different looking cache entries
(URLParserStart<Resource.hashcode>). Now the hashcode is taken from
getResolvedTemplate.