QA-9493: Refactor after review
Remove acquiredSemaphore ThreadLocal and use simple Boolean in request,
as actually the topFragment finalKey is not really needed and also risky
if keys do not match in finalize.
Do not catch JahiaServiceUnavailableExcption in in RenderChain as it
changes behaviour, rather continue returning page with missing
fragments.
Fix failing unit tests.