INDIGO-171
Work in progress.
Edit Engine => Styling according to proper design proposals for Indigo.
Edit Engine => Lots of form updates
Edit Engine => Checkbox used for toggling Panels has been redesigned as
a proper toggle switch. This works fine in Chrome, but note that in
Firefox it will fallback to a normal checkbox.

General => Created a SASS function for the creation of buttons ( saves
a lot of time and simplifies ALOT.