INDIGO-192
Renamed all variables / placeholders and extensions according to new
simplified and more explicit naming convention.

Also converted 99% of values that haven’t been variable-ised.

Variable examples:
  $COLOUR--JAHIA_BLUE
  $ICON--ARROW_DOWN
  $FORM--TEXT_INPUT--BG--COLOUR

Change the value of a variable depending on STATE:
  $FORM--TEXT_INPUT--BG--COLOUR---FOCUS

Change the value of a variable depending on CONTEXT:
  $BUTTON--BG--COLOUR
  $BUTTON--BG--COLOUR\:FORM

PLACEHOLDERS:
Placeholders are blocks of CSS that have been extracted from the SASS
and placed in the same place. This allows editing and theming to take
place in one place - meaning we don’t have to reel through thousands of
lines of code to find what we want. It also means that the same piece
of code only needs to be written once, it is then included where
necessary when SASS compiles. Here is how they are named:

  %CSS_BLOCK--BUTTON--TEXT

Change the value of a placeholder depending on CONTEXT:

  %CSS_BLOCK--BUTTON--TEXT\:FORM