INDIGO-191
I have fixed the checkbox issue in Firefox (Just displayed as normal
check boxes).

In Chrome, the checkboxes have been styled as proper toggle switches,
with an OFF and an ON position. The Text is stored as a STRING VARIABLE
in the localised SASS files ( _labels_en.scss, _labels_fr.scss ).

Unfortunately the major modifications converting the Checkbox to a
toggle switch were done in a previous commit and therefore not
registered in this commit. Go to _components.scss line 401 to see the
code.