INDIGO-177
Changed the downward arrow to three dots for “combos” that open a
picker. Note that these combos are indistinguishable from normal
combos, the only difference being a semi-adjacent attribute “readonly”.
So please keep an eye out for any undesired side effects elsewhere …

Also fixed the position problem of twin triggers in FireFox ( when a
combo has two buttons in it, i.e. “Clear” and “Arrow” ).