INDIGO-188
Done.

Prefixing all selectors is not really the best solution as it will add
A LOT of weight to the CSS file.

Instead, I simply remove the Style Sheet when in Studio Mode, and
re-apply it when in any other mode.

Note: Firefox was a little slow in detecting the mode when the page was
first loaded - meaning that the CSS was applied temporarily. This was
fairly “moche” in studio mode because the Anthracite CSS was applied
momentarily ( a few seconds sometimes ) before being removed. So, I
have added a blue screen with the DX logo ( this can be changed ) which
is displayed during this period of time. Note, that this is only when
the page is REFRESHED; if the user then changes modes with the mode
menu this is not the case.