INDIGO-68
Please use Chrome to test the Search panel in the file picker. There is
a slight bug that needs fixing in Firefox.

>>> Done
* File highlights now work in all three modes (list, display, thumbs).

* Added three dots button to files - allowing a left click upon it to
open the context menu.

* The preview button is displayed next to the three dots button.

* The three dots button and preview button are displayed when hovering
files and can be clicked without selecting the file first.

* Removed label from filter field (in details / thumbs view) and added
a magnifying glass to the field.

* The list of file sources have been converted from a tab list to a
combo. This now works in Firefox.

* Updated the style of the preview modal.
* The preview modal now displays the name of the file being previewed.
* Changed the icons in the preview window to be more coherent with the
rest of the site.

* Search panel now works 100% in Chrome.
* Search panel appears as a slide in panel.
* Search results are differentiated by a darker background.
* List of files are removed when the search panel is opened.
* When closing the search panel the list of files is automatically
reloaded (by programatically “hitting” the refresh button twice. A
500ms pause is required otherwise it messes up (I think it is because
GWT uses an animation to show / hide these files).

>>> Still to Do:
* Put refresh button in right place.
* Put icons for display mode at top of screen.
* FIREFOX ONLY: Search panel works the first time it is opened and
closed, but fails the second time.

>>> Could be improved when we have more time ( at a later date ):
* Remove the 500ms pause when refreshing the file list after closing
the search panel.
* Remove the preview icon when no-longer hovering the file list.
* When switching file source, automatically click on the first folder
in the left panel so that the file list on the right reflects the tree
on the left (this is general DX problem and NOT specific to Anthracite).