INDIGO-166
* The Side Panel is now closed when the user clicks anywhere outside
the Side Panel - even if they click on the page being edited inside the
iframe.

* Also, when the side panel is opened and the user clicks on a node,
the node is NO LONGER selected, which is good.

* The page can still be scrolled when the side panel is opened, meaning
the user can scroll to any part that they may wish - practical for drag
and drop.

* Once the user starts a drag and drop action the side panel is removed
and the user can drop their stuff into the page.