INDIGO-232 / INDIGO-56
INDIGO-232:
Created a helper called Dex for listening to the DOM. Event Handlers
can be attached to DOM elements that are added and removed, when
attributes are changed and when a GWT tree is updated.

Created a function to get around not being able to trigger clicks on
certain GWT elements. Use mouse.trigger(element, eventType).

INDIGO-56:
Done. Using the brand spanking new Dex, I have been able to fix the
settings issue ( I was having problems selecting the first item in the
settings list tree ).