INDIGO-247 / INDIGO-267
Done.

Note, @fpral , that I have kept the suggested button in the Edit
Engine. I do not find it shocking at all that there is a suggested
button in the edit engine panel and not in the workflow panel. This is
a simple fix if you don’t like it, however.

I also changed the colour from orange to blue.