INDIGO-76: Fixed
Fixed alignment issue with checkboxes.

Note that the modification is applied to ALL checkboxes. I have had a
look around the site to find other places where checkboxes are used and
there don’t seem to be any adverse affects, however, would be good to
have a double check (pardon the pun).