INDIGO-73: Fixed
The ID of the Keywords fieldset is different in French. I have changed
the CSS so that it works with the English and French ID.

It would be a good idea if we could have a class added to this fieldset
(and maybe other similar cases) that remains the same regardless of
language. @tdraier is this possible?