INDIGO-320: _new_theme.scss
Theming:

There are 6 base colours. Simply modifying these colours will suffice
in changing the theme / skin of anthracite 99% of the time:
* $color-primary:			rgb(92, 97, 100);
* $color-secondary:		rgb(250, 250, 250);
* $color-tertiary:		        rgb(0, 160, 227);
* $color-quaternary:		rgb(242, 109, 79);
* $color-quinary:			rgb(0, 160, 227);
* $color-senary:			rgb(250, 250, 250);

Variations of these base colours are automatically made with the next
section. The following settings do not need changing, however, some
base colours will require that the following is tweaked for optimal
results:

$basic-colors: (
	primary: (
		normal:	$color-primary,
		a: 		lighten($color-primary, 2%),
		b: 		lighten($color-primary, 6%),
		c: 		lighten($color-primary, 10%),
		d: 		lighten($color-primary, 25%),
		e: 		lighten($color-primary, 35%),
		f: 		darken($color-primary, 5%),
		g: 		darken($color-primary, 10%),
		h: 		rgba(darken($color-primary, 10%), .2)
	),
	secondary: (
		normal:	$color-secondary,
		a: 		darken($color-secondary, 5%),
		b: 		darken($color-secondary, 7%),
		c: 		darken($color-secondary, 10%),
		d: 		darken($color-secondary, 20%),
		e: 		darken($color-secondary, 35%)
	),
	tertiary: (
		normal:	$color-tertiary,
		a: 		lighten($color-tertiary, 2%)
	),
	quaternary: (
		normal:	$color-quaternary,
		a: 		lighten($color-quaternary, 4%)
	),
	quinary: (
		normal:	$color-quinary,
		a: 		lighten($color-quinary, 4%)
	),
	senary: (
		normal:	$color-senary,
		a: 		lighten($color-senary, 4%)
	)
);

Component color assignation:

Next, I have created a list of variables that contain colours for all
parts of the site, ie. text input colors, background colours, title
colours. For each variable there are two colours associated. One for
when the component is used on a DARK background and one when it is used
on a LIGHT background. This is NOT to be confused with a DARK and LIGHT
theme. For example, a combo box in the edit engine is displayed on a
LIGHT background, but in the edit mode it is displayed on a DARK
background. Therefore we need to specify 2 colours to be used
accordingly:

$INPUT--COLOR:(
	LIGHT: _theme-color(primary),
	DARK: _theme-color(secondary),
);

Usage:
I have created a function that will pull the value called _property(),
example:

input {
  color: _property($INPUT—COLOR, DARK);
}

Contextual colouring:
For greater and simpler control, we can replace the second parameter of
the above _property() function with a variable. That way, we can switch
the colours of a whole section from LIGHT to DARK with just one edit,
i.e:

$APP:                  DARK;
$SIDE-PANEL:    LIGHT;
$EDIT-ENGINE:  DARK;

#side-panel input {
  color: _property($INPUT—COLOR, $SIDE-PANEL);
}

#edit-engine input {
  color: _property($INPUT—COLOR, $EDIT-ENGINE);
}

The use of a variable as parameter #2 becomes interesting when
considering that the side panel may have 200 color assignations. Simply
switching the value of $SIDE-PANEL to DARK and ALL the colours in the
side panel are updated accordingly in one swift move.