BACKLOG-9709 fix issue of unpublish in all languages
The issue is related to the parent node property "j:published".
When you want to unpublish in all languages content and you unpublish
content in language "en" for example (the child node of the parent),
the property of child node "j:published" will updated to false and
also the property of parent node. Indeed, the content will not be
available for any languages.