QA-11510 Fixed StackOverflow error and added more logs
ImportExport could run into a stack overflow when computing an estimate
of the number of nodes to export. The query can end-up having a very
large number of conditions in its where clause, leading to excessive
recursions on Jackrabbit side.

From now on, such query will be splitted into several ones when a
threshold is reached (100 conditions).