QA-11759: Individual indexes can't be re-indexed once whole indexing failed
When re-indexing all Lucene indexes, JahiaSearchIndex.prepareReindexing()
is called first on each one of them. As a result, a new instance of
JahiaSecondaryIndex is created for each of them and stored in an instance
variable named "newIndex".

When a re-indexing is scheduled, scheduleReindexing() will first check
whether or not the index to process has a non-null "newIndex", and abort
processing if it does.

If a re-indexing of the whole repository fails for one of the indexes,
then re-indexation will be aborted for the remaining ones. In that case,
the value of "newIndex" has to be reset in all those remaining indexes,
otherwise any further re-indexing of those indexes will be prevented.