BACKLOG-10499: Stop failing when target and current read-only mode are equal
Until now one could not switch read-only mode to ENABLED if it was already
enabled on the current node, nor switch mode to DISABLED if it was already
disabled. On a clustered environment this was preventing from being able
to trigger a cluster-wide switch when the status of nodes were inconsistent
across the cluster.

From now on, this will no longer end-up with an error. The switch will
just be silently ignored for the current node (which is already in the
expected state).