BACKLOG-11796: Downgraded commons-fileupload from 1.4 to 1.3.3 (#321)
Use of commons-fileupload 1.4 is preventing remotepublish to work properly.

1.3.3 was the version used here before merging the Karaf upgrade branch. I cannot find any relevant information about why this library was upgraded here but nowhere else.

remotepublish module is consuming an instance of a class from commons-fileupload, which is created and provided by core, and then transmitted to remotepublish through the servlet request. Having core and remotepublish using a different version of that library leads to a ClassCastException in remotepublish's ReplayAction. The consequence of the upgrade of commons-fileupload here was that this was the one selected to be wired with remotepublish, leading to having this module using a different version that the one used by core.

I am not quite sure about the impact of downgrading the version of commons-fileupload. If this were to have any impact we could consider upgrading all versions to 1.4 instead.