BACKLOG-12803: Fix site language from authoringApi never being updated (#429)
switchLanguage(...) and switchSite(...) from authoringApi where never actually updating the current site language. Those methods are called from jContent's language and site switchers to sync the current language in the authoring API with the one selected in jContent. Not having both values in sync does lead to inconsistency when opening the publication workflow.