QABACKLOG-555: Add OWASP vulnerable dependency check
* adapt CircleCI yaml file to respect how caching works on CircleCI
* Dynamic cache key creation refactored to be multi module capable
* Try to fill Maven repository after build with all needed dependencies
for future steps before caching (cache cannot be updated later and is
then held for max 15 days or if something changes in pom.xml files)
* Do not fallback to caches, which do not have same name, if there is a
following job step to store the cache, because we want to prevent an
ever growing cache. This will cause the first build for a new cache take
longer. If its too long, we may want to change it back to the fallback
feature and find another way to clean old artifacts before saving.
* Sonar job is no longer caching Maven repository, but the Sonar/OWASP
validation data is stored in own directories, which get cached
* There are 2 different Sonar data cache keys, because OWASP dependency
check for vulnerable libraries is not done on PR, but only on merges
* Changed Maven goal on build job from install to package, because
install would only copy current module's artifact to the local Maven
repository, which gets cached for 15 days. So on cache restores in later
CircleCI workflows the newly built artifact would get overwritten in the
local repository by the old cached one.
It could be that we will have to set it back to the install goal if some
future steps require that. Need to be tested on CircleCI first.