BACKLOG-21996: Move Template resolution logic so that it can be extended from modules  (#1824)
* BACKLOG-21996 Content templates PoC
In this PoC the main idea is to make the RenderService.resolveTemplate method that is used in different parts of the code, especially in the search engine results, to become pluggeable. The current implementation only supports JCR-stored templates, but in this PoC we make that resolution pluggeable, so that we may plugin new implementation, such as an NPM-based template resolution.

This is similar to what the script resolver architecture does. All the existing template resolution was therefore moved to a new JCRTemplateResolver class.

It is expected that implementation providing their own template resolvers will register them this way, which is (currently) how we register additional script resolvers :

public class MyTemplateResolverImplements TemplateResolver {

    @Activate
    public void activate(BundleContext context) {
        // we use a copy of the list to avoid multi-threading issues
        List<TemplateResolver> templateResolvers = new ArrayList<>(renderService.getTemplateResolvers());
        templateResolvers.add(0, this);
        renderService.setTemplateResolvers(templateResolvers);
    }

    @Deactivate
    public void deactivate(BundleContext context) {
        // we use a copy of the list to avoid multi-threading issues
        List<TemplateResolver> templateResolvers = new ArrayList<>(renderService.getTemplateResolvers());
        templateResolvers.remove(this);
        renderService.setTemplateResolvers(templateResolvers);
    }

    public Template resolveTemplate(Resource resource, RenderContext renderContext) throws RepositoryException {}

    public void flushCache(String modulePath) {}

}

Note: it might be a good idea to simplify this, but beware of multi-thread issues.

* BACKLOG-21996: finalize work on making template system extensible

---------

Co-authored-by: Kevan <kevan@jahia.com>